+++ src/main/java/com/braintreegateway/PaymentMethodThreeDSecurePassThruRequest.java
@@ -1,5 +1,7 @@
 package com.braintreegateway;
 
+import com.braintreegateway.enums.ThreeDSecurePassThruNetwork;
+
 public class PaymentMethodThreeDSecurePassThruRequest extends ThreeDSecurePassThruRequest {
     private PaymentMethodRequest parent;
 
@@ -56,6 +58,12 @@ public class PaymentMethodThreeDSecurePassThruRequest extends ThreeDSecurePassTh
     @Override
     public PaymentMethodThreeDSecurePassThruRequest cavvAlgorithm(String cavvAlgorithm) {
         super.cavvAlgorithm(cavvAlgorithm);
+        return this;
+    }
+
+    @Override
+    public PaymentMethodThreeDSecurePassThruRequest network(ThreeDSecurePassThruNetwork network) {
+        super.network(network);
         return this;
     }
 }