+++ src/test/java/com/braintreegateway/integrationtest/BraintreeGatewayIT.java
@@ -41,7 +41,6 @@ public class BraintreeGatewayIT extends IntegrationTest {
         assertEquals("https://api.braintreegateway.com:443/merchants/production_merchant_id", configuration.getBaseURL() + configuration.getMerchantPath());
     }
 
-    @Disabled("Skipping pending investigation on why this fails in CI only")
     @Test
     public void tokenizeRawCreditCardDetailsWithGraphQL() {
         String query = "mutation ExampleServerSideSingleUseToken($input: TokenizeCreditCardInput!) {" +